Move locale to root
